[core] Move ColorRampRenderData to separate files
