+++ src/core-files.json
@@ -109,6 +109,7 @@
         "src/mbgl/renderer/buckets/line_bucket.cpp",
         "src/mbgl/renderer/buckets/raster_bucket.cpp",
         "src/mbgl/renderer/buckets/symbol_bucket.cpp",
+        "src/mbgl/renderer/color_ramp_render_data.cpp",
         "src/mbgl/renderer/cross_faded_property_evaluator.cpp",
         "src/mbgl/renderer/group_by_layout.cpp",
         "src/mbgl/renderer/image_atlas.cpp",
@@ -622,6 +623,7 @@
         "mbgl/renderer/buckets/line_bucket.hpp": "src/mbgl/renderer/buckets/line_bucket.hpp",
         "mbgl/renderer/buckets/raster_bucket.hpp": "src/mbgl/renderer/buckets/raster_bucket.hpp",
         "mbgl/renderer/buckets/symbol_bucket.hpp": "src/mbgl/renderer/buckets/symbol_bucket.hpp",
+        "mbgl/renderer/color_ramp_render_data.hpp": "src/mbgl/renderer/color_ramp_render_data.hpp",
         "mbgl/renderer/cross_faded_property_evaluator.hpp": "src/mbgl/renderer/cross_faded_property_evaluator.hpp",
         "mbgl/renderer/data_driven_property_evaluator.hpp": "src/mbgl/renderer/data_driven_property_evaluator.hpp",
         "mbgl/renderer/group_by_layout.hpp": "src/mbgl/renderer/group_by_layout.hpp",
