+++ src/mbgl/renderer/color_ramp_render_data.hpp
@@ -1,0 +1,26 @@
+#pragma once
+
+#include <mbgl/gfx/texture.hpp>
+#include <mbgl/util/image.hpp>
+#include <mbgl/util/optional.hpp>
+
+namespace mbgl {
+
+namespace style {
+class ColorRampPropertyValue;
+} // namespace style
+
+namespace gfx {
+class UploadPass;
+} // namespace gfx
+
+class ColorRampRenderData {
+public:
+    explicit ColorRampRenderData(const style::ColorRampPropertyValue&);
+    ~ColorRampRenderData();
+    void upload(gfx::UploadPass&);
+    const PremultipliedImage image;
+    optional<gfx::Texture> texture;
+};
+
+} // namespace mbgl
